requirements.txt
numpy~=1.26.4
pillow~=10.2.0
packaging~=24.0
ftfy~=6.2.0
regex~=2023.12.25
torchvision~=0.16.2
tqdm~=4.66.2
pandas~=2.1.4
opencv-python~=4.9.0.80
kornia~=0.6.9
scikit-learn~=1.4.1.post1
scipy~=1.12.0